refactor(web): move each page layout to its own module
Page held both layouts inline, so reading how a page is assembled meant
skipping over the other variant. Each layout now lives next to Page and
owns the props it renders, and Page composes the two variants from them.

No behaviour change: pages keep talking only to Page.
